fixed unique id issues with box, moved box to component folder
