+++ src/pages/Home/Installation.tsx
@@ -69,7 +69,7 @@ const Installation: React.FC<Props> = ({ setOpenSignupModal }) => {
                 <p>Register a NextView account here:</p>
                 <input
                   className='focus:shadow-outline my-3 w-full flex-1 appearance-none rounded border px-3 py-2 leading-tight text-gray-700 focus:outline-none  sm:w-auto sm:rounded-r-none'
-                  id='username'
+                  // id='username'
                   type='text'
                   placeholder='jsmith@example.com'
                   onChange={handleUsernameChange}
